add repo-alias flag to cluster deploy
